Move util functions to format-time file
